refactor: move cache store
[no ci]
